+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fix a bug: unable to set date formats during installation
 - Fix a bug: unable to cancel the upgrade before it begins
 - Fix a bug: in the admin calendar, the trainings' info panel shows "duration: null minutes"
+- Fix a bug: on the subscriptions page, not logged-in users do not see the action button
 - `SUPERADMIN_EMAIL` renamed to `ADMINSYS_EMAIL`
 - `scripts/run-tests.sh` renamed to `scripts/tests.sh`
 - [BREAKING CHANGE] GET `open_api/v1/invoices` won't return `stp_invoice_id` OR `stp_payment_intent_id` anymore. The new field `payment_gateway_object` will contain some similar data if the invoice was paid online by card.